test: 회원 서비스 테스트 코드 수정 및 추가
